+++ modules/.submodules.json
@@ -2,8 +2,7 @@
   "userId": [
     "digiTrustIdSystem",
     "id5IdSystem",
-    "criteortusIdSystem",
-    "liveIntentIdSystem"
+    "criteortusIdSystem"
   ],
   "adpod": [
     "freeWheelAdserverVideo",
