+++ modules/userId/index.js
@@ -81,7 +81,6 @@ import {gdprDataHandler} from '../../src/adapterManager';
 import CONSTANTS from '../../src/constants.json';
 import {module} from '../../src/hook';
 import {unifiedIdSubmodule} from './unifiedIdSystem.js';
-import {liveIntentIdSubmodule} from './liveIntentIdSystem.js';
 import {pubCommonIdSubmodule} from './pubCommonIdSystem.js';
 
 const MODULE_NAME = 'User ID';
@@ -485,6 +484,5 @@ init(config);
 // add submodules after init has been called
 attachIdSystem(pubCommonIdSubmodule);
 attachIdSystem(unifiedIdSubmodule);
-attachIdSystem(liveIntentIdSubmodule);
 
 module('userId', attachIdSystem);
